Add simple chatbot demo, prepare enhancement for README
The demo is a simple chat interface running on flask. I will add a docker image for easy deployment as well.
For the readme, I would like to be able to click on different commands instead of pasting them all after each other, as the commands are very similar to each other